Updating references to frame number / position to long int, there were still quite a few old "int" declarations, which limits the length of frame number. Also, updated precision of KeyFrames to use double (instead of float) for higher precision, which fixed lots of issues with very long videos (since our FrameMapper used a KeyFrame object to create a map to new frames)

+++ include/AudioReaderSource.h
@@ -64,7 +64,7 @@ namespace openshot
 		int64 original_frame_number; /// The current frame to read from
 		int64 frame_number; /// The current frame number
 		tr1::shared_ptr<Frame> frame; /// The current frame object that is being read
-		int frame_position; /// The position of the current frame's buffer
+		long int frame_position; /// The position of the current frame's buffer
 		double estimated_frame; /// The estimated frame position of the currently playing buffer
 		int estimated_samples_per_frame; /// The estimated samples per frame of video
 
@@ -121,7 +121,7 @@ namespace openshot
 	    tr1::shared_ptr<Frame> getFrame() const { return frame; }
 
 	    /// Get the estimate frame that is playing at this moment
-	    int getEstimatedFrame() const { return int(estimated_frame); }
+	    long int getEstimatedFrame() const { return long(estimated_frame); }
 
 	    /// Set Speed (The speed and direction to playback a reader (1=normal, 2=fast, 3=faster, -1=rewind, etc...)
 	    void setSpeed(int new_speed) { speed = new_speed; }
