+++ include/Coordinate.h
@@ -55,11 +55,11 @@ namespace openshot {
 	private:
 		bool increasing; ///< Is the Y value increasing or decreasing?
 		Fraction repeated; ///< Fraction of repeated Y values (for example, 1/3 would be the first Y value of 3 repeated values)
-		float delta; ///< This difference in Y value (from the previous unique Y value)
+		double delta; ///< This difference in Y value (from the previous unique Y value)
 
 	public:
-		float X; ///< The X value of the coordinate (usually representing the frame #)
-		float Y; ///< The Y value of the coordinate (usually representing the value of the property being animated)
+		double X; ///< The X value of the coordinate (usually representing the frame #)
+		double Y; ///< The Y value of the coordinate (usually representing the value of the property being animated)
 
 		/// The default constructor, which defaults to (0,0)
 		Coordinate();
@@ -67,7 +67,7 @@ namespace openshot {
 		/// @brief Constructor which also sets the X and Y
 		/// @param x The X coordinate (usually representing the frame #)
 		/// @param y The Y coordinate (usually representing the value of the property being animated)
-		Coordinate(float x, float y);
+		Coordinate(double x, double y);
 
 		/// @brief Set the repeating Fraction (used internally on the timeline, to track changes to coordinates)
 		/// @param is_repeated The fraction representing how many times this coordinate Y value repeats (only used on the timeline)
@@ -85,7 +85,7 @@ namespace openshot {
 
 		/// @brief Set the delta / difference between previous coordinate value (used internally on the timeline, to track changes to coordinates)
 		/// @param new_delta Indicates how much this Y value differs from the previous Y value
-		void Delta(float new_delta) { delta=new_delta; }
+		void Delta(double new_delta) { delta=new_delta; }
 
 		/// Get the delta / difference between previous coordinate value (used internally on the timeline, to track changes to coordinates)
 		float Delta() { return delta; }
